(PC-19440) fix(statusBar): add useWhiteStatusBarWithoutReactNavigation
